projects/06-smartway/backend/requirements.txt
# FastAPI Backend
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# LangGraph & LangChain
langgraph==0.2.0
langchain-openai==0.1.0
langchain-core==0.1.0

# Additional
python-multipart==0.0.6
